new Math CYU

+++ trapezoid.py
@@ -5,9 +5,9 @@
 bottom = input ("Enter the long base length ")
 top = input ("Enter the short base length ")
 height = input ("Enter the altitude ")
-bottom = int(bottom)
-top = int(top)
-height = int(height)
+bottom = float(bottom)
+top = float(top)
+height = float(height)
 
 
 area = (bottom + top) / 2 * height
